[Feat]: upload user name - 로그인, 회원가입 시 유저정보 캐싱과 서버 업로드 기능 구현

+++ app/src/main/java/com/hapataka/questwalk/data/model/UserModel.kt
@@ -4,9 +4,9 @@ data class UserModel(
     val userId: String,
     var nickName: String = "",
     var characterId: Int = 1,
-    var totalTime: Int = 0,
+    var totalTime: Long = 0L,
     var totalDistance: Float = 0f,
-    var totalStep: Long = 0,
+    var totalStep: Long = 0L,
 ) {
     fun changeNickName(newNickName: String) {
         this.nickName = newNickName
@@ -16,7 +16,7 @@ data class UserModel(
         this.characterId = newCharacterId
     }
 
-    fun updateTotalInfo(time: Int, distance: Float, step: Long) {
+    fun updateTotalInfo(time: Long, distance: Float, step: Long) {
         this.totalTime += time
         this.totalDistance += distance
         this.totalStep += step
